refactor: select query 단일 클래스로 관리
